Добавляет загрузку комментариев
